requirements.txt
affinegap==1.10
appdirs==1.4.3
backports.csv==1.0.5
BTrees==4.4.1
canonicalize==1.3
categorical-distance==1.9
click==6.7
dedupe==1.6.13
dedupe-hcluster==0.3.2
DoubleMetaphone==0.1
-e git+https://github.com/hammady/python-fastcluster@b676a2d3da4bb6c9d265cd268c9ecd8173a63360#egg=fastcluster
future==0.16.0
haversine==0.4.5
highered==0.2.1
Levenshtein-search==1.4.4
numpy==1.12.1
packaging==16.8
persistent==4.2.4.2
psycopg2==2.7.1
pyhacrf-datamade==0.2.1
PyLBFGS==0.2.0.5
pyparsing==2.2.0
python-dateutil==2.6.0
python-dotenv==0.6.4
PyYAML==3.12
-e git+https://github.com/rayyanqcri/rayyan-dedup@33bdd6a82a4b60d4005a226260084b96611ed37b#egg=rayyan_dedup
rlr==2.4.3
simplecosine==1.1
simplejson==3.10.0
six==1.10.0
Unidecode==0.4.20
zope.index==4.3.0
zope.interface==4.4.1
